Moved extras/datatypes to extras/interleaved and moved in InterleavedComputation
