+++ examples/jobs/src/main/scala/com/example/jobs/Bob.scala
@@ -3,7 +3,7 @@ package com.example.jobs
 import indigo._
 import indigoextras.jobs._
 import indigo.shared.EqualTo._
-import indigoextras.datatypes.IncreaseTo
+import indigoextras.interleaved.IncreaseTo
 
 // Bob is our NPC 'Actor'
 final case class Bob(position: Point, workSchedule: WorkSchedule[Bob, Unit], state: BobState) {
