+++ indigo/indigo-extras/src/main/scala/indigoextras/interleaved/InterleavedComputation.scala
@@ -1,4 +1,4 @@
-package indigoextras.ui
+package indigoextras.interleaved
 
 import indigo.shared.time.GameTime
 
@@ -8,7 +8,7 @@ import scala.annotation.tailrec
   * The interface for one step in an interleaved computation.
   *
   * @tparam ReferenceData Type of the static data made available to every step in the computation
-  * @tparam ResultModel Type of the interleaved computation's (partial/complete) result
+  * @tparam ResultModel   Type of the interleaved computation's (partial/complete) result
   */
 trait MonitoredStep[ReferenceData, ResultModel] {
 
@@ -24,7 +24,7 @@ trait MonitoredStep[ReferenceData, ResultModel] {
     * Perform this computation step.
     *
     * @param reference The static reference data available to all steps
-    * @param current The running result produced by the computation steps prior to this one
+    * @param current   The running result produced by the computation steps prior to this one
     * @return The result state generated by this step
     */
   def perform(reference: ReferenceData, current: ResultModel): ResultModel
@@ -34,13 +34,13 @@ trait MonitoredStep[ReferenceData, ResultModel] {
   * Represents a computation whose steps are to be performed within frames, returning control
   * back to Indigo at the end of each frame to allow graphics and other elements to be updated.
   *
-  * @param currentResult The result data object before the (remaining) steps are performed
-  * @param referenceData Static data available for reference by all computation steps
-  * @param steps The remaining computation steps
+  * @param currentResult      The result data object before the (remaining) steps are performed
+  * @param referenceData      Static data available for reference by all computation steps
+  * @param steps              The remaining computation steps
   * @param averageTimePerUnit The average time taken per computation step unit size
-  * @param sizeCompleted The number of computation step units completed so far
+  * @param sizeCompleted      The number of computation step units completed so far
   * @tparam ReferenceData Type of the static data made available to every step in the computation
-  * @tparam ResultModel Type of the interleaved computation's (partial/complete) result
+  * @tparam ResultModel   Type of the interleaved computation's (partial/complete) result
   */
 final case class InterleavedComputation[ReferenceData, ResultModel](
     currentResult: ResultModel,
@@ -77,8 +77,9 @@ final case class InterleavedComputation[ReferenceData, ResultModel](
 
   /**
     * Recursive helper function to perform at least one step, and then more if there is still time.
+    *
     * @param frameDuration The duration of the frame
-    * @param taken How long has been taken so far
+    * @param taken         How long has been taken so far
     * @return The new interleaved computation state after performing more steps
     */
   @tailrec
