+++ indigo/indigo-extras/src/test/scala/indigoextras/datatypes/TimeVaryingValueTests.scala
@@ -1,4 +1,4 @@
-package indigoextras.datatypes
+package indigoextras.interleaved
 
 import utest._
